RENAME - CustomAttribute.py -> MyAttribute.py
